Implement multi-root support
